Code cleanup

+++ code/src/sputil/core/StopWatch.cpp
@@ -48,5 +48,6 @@ double StopWatch::seconds() const
 double StopWatch::milliseconds() const
 {
     constexpr double microsecondsInMillisecond = 1000.0;
-    return chrono::duration_cast<chrono::microseconds>(m_ended - m_started).count() / microsecondsInMillisecond;
+    return (double) chrono::duration_cast<chrono::microseconds>(m_ended - m_started).count() /
+           microsecondsInMillisecond;
 }
